fix cyclic dep
