+++ frontend/javascripts/viewer/geometries/materials/plane_material_factory.ts
@@ -19,10 +19,10 @@ import {
   getMagInfoByLayer,
   getMappingInfoForSupportedLayer,
   getSegmentationLayerWithMappingSupport,
-  getTransformedVoxelSize,
   getVisibleSegmentationLayer,
 } from "viewer/model/accessors/dataset_accessor";
 import {
+  getTransformedVoxelSize,
   getTransformsForLayer,
   getTransformsPerLayer,
   invertAndTranspose,
