+++ frontend/javascripts/viewer/model/helpers/nml_helpers.ts
@@ -33,8 +33,10 @@ import type {
   WebknossosState,
 } from "viewer/store";
 import { findGroup } from "viewer/view/right-border-tabs/trees_tab/tree_hierarchy_view_helpers";
-import { getTransformedVoxelSize } from "../accessors/dataset_accessor";
-import { getTransformsForSkeletonLayer } from "../accessors/dataset_layer_transformation_accessor";
+import {
+  getTransformedVoxelSize,
+  getTransformsForSkeletonLayer,
+} from "../accessors/dataset_layer_transformation_accessor";
 import { getNodePosition } from "../accessors/skeletontracing_accessor";
 import { min } from "./iterator_utils";
